moved query show ica
